Config: Add botservice 2021-05-01-preview
The reason to add this is because the provider is currently using the same version for track1, but the track1 SDK is generated from a broken swagger version, which technically need to be regenerated, but they rejected to regenerate for track1 for such case. See https://github.com/Azure/azure-sdk-for-go/issues/19944.